backend/requirements.txt
# Core scraping dependencies
requests==2.31.0
beautifulsoup4==4.12.2
selenium>=4.0.0

# Database
psycopg2-binary>=2.9.9

# CLI
click==8.1.7
tabulate==0.9.0

# Configuration
python-dotenv==1.0.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
